fix: 修改回原来的Integer

+++ boot/datasource-starter/src/main/java/com/kewen/framework/datasource/context/DbCurrentUser.java
@@ -12,7 +12,7 @@ public interface DbCurrentUser {
      * 用户id
      * @return
      */
-    Long getUserId();
+    Integer getUserId();
 
     /**
      * 用户名
