+++ boot/datasource-starter/src/main/java/com/kewen/framework/datasource/plug/GlobalFillConfig.java
@@ -26,7 +26,7 @@ public class GlobalFillConfig implements MetaObjectHandler {
     public void insertFill(MetaObject metaObject) {
 
         this.strictInsertFill(metaObject, "createTime", LocalDateTime.class, LocalDateTime.now());
-        this.strictInsertFill(metaObject, "createUserId", Long.class, dbCurrentUser.getUserId());
+        this.strictInsertFill(metaObject, "createUserId", Integer.class, dbCurrentUser.getUserId());
         this.strictInsertFill(metaObject, "createUserName", String.class, dbCurrentUser.getUserName());
 
     }
